Quickfix for getting app in visible bounds
setting the default window size has changed due to the new architecture, I currently set it in MainScreenView, but there may be a better place to put this as it is still not as perfect as it was before.

Similarly, made a small quick change to the ProjectWizard to make it bearable.