update jdk/jfx
Updates the jdk from oraclejdk8 to amazon corretto jdk 11, and javafx to openjfx 11.

Tornadofx is updated to the 2.0.0rc which requires the sonatype maven repo.

JFoenix is updated to 9.0.9
This required some changes in the snackbar to adhere to a constructor change.

Javafx imports also had to be moved in the StatusIndicator.